Delete key_generator_test.rb -> add tests to offset_test. In process of altering tests
